Add extension to links to posts
Otherwise, we get False matches for the extension further down the
pipeline, when replacing the original extension with `.html`.
For example, a link [[R. Smith]] would result in `. smith` as a detected
extension. Rather, we want `R. Smith.md`, so that we can slugify `R.
Smith` and swap `.md` with `.html`.