:construction: Generate archive page.
There seems to be a bug with getting the creation date from the
frontmatter. Also, getting the creation date from git seems to fail.
Furthermore, getting the creation date is very slow, so it should
probably be cached in the note metadata.

At this point, the `Note` struct is becoming a bit large, so it may make
sense to split it into a `metadata` and a `content` struct or something.
Needs more thought.