Add brackets to placeholders
This helps the detection of header labels for instance.
Suppose we had a link to a blockref, something like this.
```
[[#^someid]]
```
This would get detected by the blockref detector, which is why we
first remove the links.
When a link is removed, it is replaced by a placeholder of the form
```
{{ <hash> }}
```
Before, it was just the hash. Because the hash contains only regular
`\w` characters, `<hash>^id` would not detect `^id` as a blockref label,
even though it should, because originally, it was `[[]]^id`, so it is
preceded by a `]`, which is not a `\w` character.